+++ fltk-derive/src/widget.rs
@@ -615,7 +615,6 @@ pub fn impl_widget_trait(ast: &DeriveInput) -> TokenStream {
                             0
                         }
                     }
-                    let _old_data = self.handle_data();
                     let a: *mut Box<dyn FnMut(Event) -> bool> = Box::into_raw(Box::new(Box::new(cb)));
                     let data: *mut raw::c_void = a as *mut raw::c_void;
                     let callback: custom_handler_callback = Some(shim);
@@ -641,7 +640,6 @@ pub fn impl_widget_trait(ast: &DeriveInput) -> TokenStream {
                             0
                         }
                     }
-                    let _old_data = self.handle_data();
                     let a: *mut Box<dyn FnMut(&mut Self, Event) -> bool> = Box::into_raw(Box::new(Box::new(cb)));
                     let data: *mut raw::c_void = a as *mut raw::c_void;
                     let callback: custom_handler_callback2 = Some(shim);