+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - <image type>Image::write_to_file has an AsRef<std::path::Path> parameter.
 - Add SurfaceDevice trait and ImageSurface type.
 - Prelim support for musl builds.
+- Remove back-compat SurfaceDevice::set_current().
 
 ## [0.10.1] - 2020-10-30
 ### Changes